change structure of bedgraphLike
